rag/requirements.txt
# API
fastapi==0.129.0
uvicorn==0.41.0
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
slowapi==0.1.9

# LangChain RAG
langchain==1.2.10
langchain-core==1.2.14
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-chroma==1.1.0

# LLM providers (keep only what you USE)
langchain-google-genai==4.2.1     # if using Gemini

# Vector DB
chromadb==1.5.0

# Document loaders
pypdf==6.7.1
python-docx==1.2.0
beautifulsoup4==4.14.3

# Database (if using MongoDB)
pymongo==4.16.0
motor==3.7.1

# Utilities
tiktoken==0.12.0
httpx==0.28.1